+++ src/viewport/ViewportWidget.hpp
@@ -31,7 +31,7 @@ protected:
 
 private:
     double widgetPixelRatio() const;
-    glm::vec2 mapToRenderer(const QPoint& p);
+    glm::dvec2 mapToRenderer(const QPoint& p);
 
     SP<ViewportRenderer> _renderer;
     CameraController _cameraController;
